split lexer into multiple files
