+++ internal/semantic/image_tokens.go
