+++ internal/semantic/image_tokens_test.go
