+++ internal/semantic/network_tokens.go
@@ -1,12 +1,14 @@
 package semantic
 
 import (
+	"strings"
+
 	"github.com/onlyati/quadlet-lsp/internal/utils"
 	protocol "github.com/tliron/glsp/protocol_3_16"
 )
 
-// readPodValue parses pod value like: 'foo.pod'
-func (l *lexer) readPodValue() {
+// readNetworkValue parses network value like: 'foo.network'
+func (l *lexer) readNetworkValue() {
 	continueDetected := false
 	for {
 		l.skipInlineWhitespace()
@@ -25,7 +27,11 @@ func (l *lexer) readPodValue() {
 			return
 		default:
 			if utils.IsLetter(l.ch) || l.ch == '/' {
-				token := l.readUntil(map[rune]struct{}{}, string(protocol.SemanticTokenTypeParameter))
+				token := l.readUntil(map[rune]struct{}{}, string(protocol.SemanticTokenTypeString))
+				if strings.HasSuffix(strings.TrimSpace(token.text), ".network") {
+					token.tokenType = string(protocol.SemanticTokenTypeParameter)
+				}
+
 				l.queue = append(l.queue, token)
 			} else {
 				l.readRune() // Avoid infinite loop on unkown field
