+++ internal/semantic/network_tokens_test.go
