+++ internal/semantic/semantic.go
@@ -3,9 +3,6 @@
 package semantic
 
 import (
-	"unicode/utf8"
-
-	"github.com/onlyati/quadlet-lsp/internal/utils"
 	protocol "github.com/tliron/glsp/protocol_3_16"
 )
 
@@ -66,257 +63,6 @@ func CalculateSemanticTokens(fileText string) (*protocol.SemanticTokens, error) 
 	}
 
 	return &protocol.SemanticTokens{Data: data}, nil
-}
-
-type token struct {
-	line      protocol.UInteger
-	charPos   protocol.UInteger
-	length    protocol.UInteger
-	tokenType string
-	text      string
-}
-
-type lexer struct {
-	input        string
-	position     protocol.UInteger
-	readPosition protocol.UInteger // Next character after position
-	lineNumber   protocol.UInteger
-	lineStart    protocol.UInteger
-	ch           rune
-	queue        []token
-}
-
-func newLexer(input string) *lexer {
-	l := &lexer{
-		input:        input,
-		position:     0,
-		readPosition: 0,
-		lineNumber:   0,
-		lineStart:    0,
-		ch:           0,
-		queue:        []token{},
-	}
-	l.readRune()
-	return l
-}
-
-func (l *lexer) readRune() {
-	if l.readPosition >= uint32(len(l.input)) {
-		l.ch = 0 // EOF
-		l.position = l.readPosition
-		return
-	}
-
-	r, width := utf8.DecodeRuneInString(l.input[l.readPosition:])
-
-	l.ch = r
-	l.position = l.readPosition
-	l.readPosition += uint32(width)
-}
-
-// func (l *lexer) peekRune() rune {
-// 	if l.readPosition >= uint32(len(l.input)) {
-// 		return 0
-// 	}
-// 	r, _ := utf8.DecodeRuneInString(l.input[l.readPosition:])
-// 	return r
-// }
-
-func (l *lexer) handleNewLine() {
-	l.lineNumber++
-	l.readRune()
-	l.lineStart = l.position
-}
-
-func (l *lexer) skipInlineWhitespace() {
-	for l.ch == ' ' || l.ch == '\t' || l.ch == '\r' {
-		l.readRune()
-	}
-}
-
-func (l *lexer) skipWhitespace() {
-	for l.ch == ' ' || l.ch == '\t' || l.ch == '\n' || l.ch == '\r' {
-		if l.ch == '\n' {
-			l.handleNewLine()
-			continue
-		}
-		l.readRune()
-	}
-}
-
-func (l *lexer) nextToken() token {
-	// if something has been put into the queue, then empty it
-	if len(l.queue) > 0 {
-		tok := l.queue[0]
-		l.queue = l.queue[1:]
-		return tok
-	}
-
-	for {
-		l.skipWhitespace()
-
-		switch l.ch {
-		case '#':
-			return l.readComment()
-		case '[':
-			return l.readSection()
-		case '\\':
-			return l.readOperator()
-		case 0:
-			return token{
-				tokenType: "eof",
-				line:      l.lineNumber,
-				charPos:   l.position,
-				length:    0,
-				text:      "",
-			}
-		default:
-			if utils.IsLetter(l.ch) {
-				l.readAssignment()
-				return l.nextToken()
-			}
-
-			// This is something unknow, just iterate over it until we find
-			// something interesting
-			l.readRune()
-			continue
-		}
-	}
-}
-
-func (l *lexer) readOperator() token {
-	startByte := l.position
-	charPos := utils.Utf16Len(l.input[l.lineStart:l.position]) // Calc column in UTF-16
-
-	l.readRune()
-
-	// Measure the content we just read in UTF-16 units
-	sectionText := l.input[startByte:l.position]
-
-	return token{
-		line:      l.lineNumber,
-		charPos:   charPos,
-		length:    utils.Utf16Len(sectionText),
-		tokenType: string(protocol.SemanticTokenTypeOperator),
-		text:      sectionText,
-	}
-}
-
-func (l *lexer) readUntil(delimiters map[rune]struct{}, tokenType string) token {
-	startByte := l.position
-	charPos := utils.Utf16Len(l.input[l.lineStart:l.position]) // Calc column in UTF-16
-
-	// Append default delimiters
-	delimiters['\n'] = struct{}{}
-	delimiters['\\'] = struct{}{}
-	delimiters[0] = struct{}{}
-
-	inList := func(ch rune) bool {
-		_, ok := delimiters[ch]
-		return ok
-	}
-
-	for !inList(l.ch) {
-		l.readRune()
-	}
-
-	// Measure the content we just read in UTF-16 units
-	text := l.input[startByte:l.position]
-
-	return token{
-		line:      l.lineNumber,
-		charPos:   charPos,
-		length:    utils.Utf16Len(text),
-		tokenType: tokenType,
-		text:      text,
-	}
-}
-
-func (l *lexer) readAssignment() {
-	startByte := l.position
-	charPos := utils.Utf16Len(l.input[l.lineStart:l.position]) // Calc column in UTF-16
-
-	for l.ch != '=' && l.ch != '\n' && l.ch != '\\' && l.ch != 0 {
-		l.readRune()
-	}
-
-	// Measure the content we just read in UTF-16 units
-	propText := l.input[startByte:l.position]
-
-	// If we hit a \n, it means it was a value line, if we hit '=' it is property
-	if l.ch == '\n' || l.ch == '\\' || l.ch == 0 {
-		l.queue = append(l.queue, token{
-			line:      l.lineNumber,
-			charPos:   charPos,
-			length:    utils.Utf16Len(propText),
-			tokenType: string(protocol.SemanticTokenTypeString),
-			text:      propText,
-		})
-		return
-	}
-
-	// We hit '=' a sign, put property to queue, then analyze line further
-	l.queue = append(l.queue, token{
-		line:      l.lineNumber,
-		charPos:   charPos,
-		length:    utils.Utf16Len(propText),
-		tokenType: string(protocol.SemanticTokenTypeKeyword),
-		text:      propText,
-	})
-
-	if l.ch == '=' {
-		l.queue = append(l.queue, l.readOperator())
-
-		// Check if we had any special parse for property, if not just read value
-		if fn, ok := specialFunctionMap[propText]; ok {
-			fn(l)
-		} else {
-			l.queue = append(l.queue, l.readUntil(map[rune]struct{}{}, string(protocol.SemanticTokenTypeString)))
-		}
-	}
-}
-
-func (l *lexer) readSection() token {
-	startByte := l.position
-	charPos := utils.Utf16Len(l.input[l.lineStart:l.position]) // Calc column in UTF-16
-
-	for l.ch != ']' && l.ch != '\n' && l.ch != 0 {
-		l.readRune()
-	}
-	if l.ch == ']' {
-		l.readRune()
-	}
-
-	// Measure the content we just read in UTF-16 units
-	sectionText := l.input[startByte:l.position]
-
-	return token{
-		line:      l.lineNumber,
-		charPos:   charPos,
-		length:    utils.Utf16Len(sectionText),
-		tokenType: string(protocol.SemanticTokenTypeNamespace),
-		text:      sectionText,
-	}
-}
-
-func (l *lexer) readComment() token {
-	startByte := l.position
-	charPos := utils.Utf16Len(l.input[l.lineStart:l.position]) // Calc column in UTF-16
-
-	for l.ch != '\n' && l.ch != 0 {
-		l.readRune()
-	}
-
-	// Measure the content we just read in UTF-16 units
-	commentText := l.input[startByte:l.position]
-
-	return token{
-		line:      l.lineNumber,
-		charPos:   charPos,
-		length:    utils.Utf16Len(commentText),
-		tokenType: string(protocol.SemanticTokenTypeComment),
-		text:      commentText,
-	}
 }
 
 func parseQuadlet(fileText string) []token {
